Retrieve saved tag from Redux for each receipt in ReceiptList
