+++ src/components/Receipt/ReceiptList/ReceiptList.js
@@ -6,6 +6,7 @@ import MenuButton from '../../UI/Button/MenuButton/MenuButton'
 import { connect } from 'react-redux'
 import * as actionTypes from '../../../store/actions/actions'
 import getAllReceipts from '../../../services/getAllReceipts'
+import getAllTags from '../../../services/getAllTags'
 const smallDevice = window.matchMedia('(max-width: 800px)').matches
 export class ReceiptList extends Component {
     state = {  
@@ -13,6 +14,7 @@ export class ReceiptList extends Component {
     }
     componentDidMount() {
         this.fetchReceipts()
+        this.fetchTags()
     }
     render() {
         return (
@@ -28,10 +30,15 @@ export class ReceiptList extends Component {
 
     fetchReceipts = async() => {
         const receipts = await getAllReceipts()
-        console.log(receipts)
         this.setState({ loaded: true })
         this.props.onReceiptsAdded(receipts)
     }
+
+    fetchTags = async() => {
+		const tags = await getAllTags()
+        this.setState({ tags })
+		this.props.onTagsAdded(tags)
+	}
 }
 export const mapStateToProps = state => {
     return {
@@ -40,7 +47,9 @@ export const mapStateToProps = state => {
 }
 export const mapDispatchToProps = dispatch => {
     return {
-        onReceiptsAdded: (receipts) => dispatch({type: actionTypes.ADD_RECEIPTS, receipts: receipts}) 
+        onReceiptsAdded: (receipts) => dispatch({type: actionTypes.ADD_RECEIPTS, receipts: receipts}),
+        onTagsAdded: (tags) => dispatch({ type: actionTypes.ADD_TAGS, tags: tags }) 
     }
 }
+
 export default connect(mapStateToProps, mapDispatchToProps)(ReceiptList)
