+++ src/components/Receipt/ReceiptView/ReceiptView.js
@@ -8,14 +8,12 @@ import BackDrop from '../../UI/BackDrop/BackDrop'
 import deleteReceipt from '../../../services/deleteReceipt'
 import receiptInput from '../../../helpers/receiptInputs'
 import SavedTagItem from '../../UI/TagItem/SavedTagItem/SavedTagItem'
-import { connect } from 'react-redux'
 const smallDevice = window.matchMedia('(max-width: 645px)').matches
 var size;
 if(smallDevice) 
   size = 'medium'
 else 
   size = 'small'
-
 
 class ReceiptView extends Component {
 
@@ -50,7 +48,7 @@ class ReceiptView extends Component {
                   <p className='receipt-font'>{receipt['description']}</p>
                 </div>
               </div>
-              <SavedTagItem name={this.getTagName()} color={this.getTagColor()}/>
+              <SavedTagItem name={this.props.tagName} color={this.props.TagColor}/>
             </div>
           </Receipt>
 
@@ -68,11 +66,6 @@ class ReceiptView extends Component {
       </Modal>
     )
   }
-
-  getTagName = () => { return this.props.tags[this.props.receipt.tag_id - 1].category }
-
-  getTagColor = () => { return this.props.tags[this.props.receipt.tag_id - 1].color }
-
   onDeleteHandler = async() => {
     let receipt_id = this.props.receiptId
     const response = await deleteReceipt(receipt_id)
@@ -80,11 +73,9 @@ class ReceiptView extends Component {
     this.props.onClosePopup()
     this.props.onGetAllReceipts()
   }
-  
   onCancelHandler = () => {
     this.setState({confirmation: false})
   }
-  
   renderConfirmationMessage = () => {
     return (
       <Fragment>
@@ -95,14 +86,7 @@ class ReceiptView extends Component {
       </Fragment>
     )
   }
+  onConfirmationTrue = () => { this.setState({confirmation: true}) }
+}
 
-  onConfirmationTrue = () => {
-    this.setState({confirmation: true})
-  }
-}
-export const mapStateToProps = state => {
-	return {
-		tags: state.tags
-	}
-}
-export default connect(mapStateToProps)(ReceiptView)
+export default ReceiptView
